requirements.txt
django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
gunicorn
whitenoise
django-filter>=23.0
python-dotenv
Pillow
